requirements.txt
flask
flask-cors
python-dotenv
google-genai==1.62.0
gunicorn
